requirements_server.txt
streamlit
streamlit_extras
pandas
geopandas
google-generativeai
sparql_dataframe
langchain
langchain_groq
groq
datacommons-pandas
datacommons
chromadb
decouple
google.generativeai
fastapi
groq
llama-index-core
llama-index-llms-openai
llama-index-embeddings-huggingface
